requirements.txt
APScheduler==3.10.4
fpdf==1.7.2
PyYAML==6.0.2
requests==2.32.3
